requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
et_xmlfile==2.0.0
fastapi==0.115.12
fonttools==4.57.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn[standard]
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
python-jose~=3.4.0
passlib~=1.7.4
python-dotenv~=1.1.0